% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_partable.R
\name{expand_mod}
\alias{expand_mod}
\title{Expand lavaan partable into different formats.}
\usage{
expand_mod(mod_block)
}
\arguments{
\item{mod_block}{d data.frame specifying lavaan partable}
}
\value{
list with named entry for each format
}
\description{
* args: array with value of each parameter
 * mod_block: original lavaan partable
 * template: lavaan syntax with labels in handlebars (e.g. "{{label_name}}")
 * mod_inst: lavaan syntax with labels replaced by their value
 * mod: lavaan syntax with unaltered labels
}

