% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lav_syntax.R
\name{group_out}
\alias{group_out}
\title{Group like-terms from left-hand side, and output lavaan syntax.}
\usage{
group_out(d)
}
\arguments{
\item{d}{data.frame specifying lavaan partable}
}
\description{
Separates three kinds of equations
 * latent variable: uses "=~" operator
 * latent covariance: uses '~~' with lhs != rhs
 * residual variance: uses '~~' with lhs == rhs
}
\examples{
# Build partable
make_meas <- make_edge(op='=~')
make_cov <-  make_edge(op='~~')
partable <- rbind(
   make_meas('A', c('A1','A2'), 'lam_A'),
   make_meas('B', c('B1','B2'), 'lam_B'),
   make_cov( 'A', 'B', 'cov_AxB'),
   make_cov( 'A', 'A', ""),
   make_cov( 'B', 'B', "")
)
# Convert to lavaan string syntax
group_out(partable)
}

