% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_templates.R
\name{inst_template}
\alias{inst_template}
\title{Fill in template using mustache syntax}
\usage{
inst_template(template, args)
}
\arguments{
\item{template}{string}

\item{args}{list (or coercible to list) used for context in expanding mustaches}
}
\description{
Fill in template using mustache syntax
}
\examples{
template <- "name: {{name}}, id: {{id}}"
inst_template(template, list(name='michael', id=1))
}

