% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_partable.R
\name{make_edge}
\alias{make_edge}
\title{Generic function to create row(s) for lavaan partable}
\usage{
make_edge(lhs, op, rhs, label, value = "")
}
\arguments{
\item{lhs}{left-hand side of model equation}

\item{op}{operator connecting both sides}

\item{rhs}{right-hand side of model equation}

\item{label}{label for parameter that weights rhs}

\item{value}{value for parameter}
}
\description{
Curried convenience function which passes arguments straight into data.frame.
Each row in the partable specifies an equation, e.g.
"lhs ~~ label*rhs", where "~~" would be the operator.
}

