% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{modind}
\alias{modind}
\alias{modification_index}
\title{Summary of Modification Indices}
\usage{
modind(fit, n = 20)
}
\arguments{
\item{fit}{A fitted object made in lavaan}

\item{n}{The first number of elements of the modification index data frame.}
}
\value{
A data frame with the modification indices.
}
\description{
A modification index data.frame sort by power.
}
\details{
This is a wrapper function.
}
\examples{
 The famous Holzinger and Swineford (1939) example
 HS.model <- " visual  =~ x1 + x2 + x3
               textual =~ x4 + x5 + x6
               speed   =~ x7 + x8 + x9 "

 fit <- cfa(HS.model, data=HolzingerSwineford1939)
 modind(fit)
}
\seealso{
\code{\link[lavaan]{modificationIndices}}
}
\keyword{internal}
