% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lav_syntax.R
\name{param}
\alias{param}
\title{Mustache param labels, if they exist}
\usage{
param(label)
}
\arguments{
\item{label}{array of parameter labels or empty strings "".}
}
\description{
Mustache param labels, if they exist
}

