% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_partable.R
\name{remove_var}
\alias{remove_var}
\title{Remove variable from lavaan partable.}
\usage{
remove_var(d, varname, expand = FALSE)
}
\arguments{
\item{d}{data.frame specifying lavaan partable}

\item{varname}{variables to be removed}

\item{expand}{should results be passed to \code{\link{expand_mod}}?}
}
\description{
Remove variable from lavaan partable.
}

