% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_templates.R
\name{rm_mustache}
\alias{rm_mustache}
\title{Remove all mustaches from string, leaving their contents}
\usage{
rm_mustache(s)
}
\arguments{
\item{s}{string, e.g. "my name is {{name}}"}
}
\description{
Remove all mustaches from string, leaving their contents
}

