% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{two_factor}
\alias{two_factor}
\title{Example two factor model}
\format{A list of lavaan models. The list has five objects, as described
in \code{\link{expand_mod}}.}
\usage{
two_factor
}
\description{
Example two factor model
}
\keyword{datasets}

