% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_templates.R
\name{uniq_mustache}
\alias{uniq_mustache}
\title{Get all unique expressions inside mustaches}
\usage{
uniq_mustache(s)
}
\arguments{
\item{s}{string}
}
\description{
Get all unique expressions inside mustaches
}
\examples{
template <- "name: {{name}}, id: {{id}}"
uniq_mustache(template)     # c('name', 'id')
}

