% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{write_model}
\alias{write_model}
\title{Write model file and instance to disk}
\usage{
write_model(bundle, model_file = "", inst_file = "")
}
\arguments{
\item{bundle}{output from \code{\link{expand_mod}}}

\item{model_file}{path to save mod string}

\item{inst_file}{path to save mod_inst string}
}
\description{
Write model file and instance to disk
}

